+++ neo/Implementations/Blockchains/LevelDB/DebugStorage.py
@@ -5,7 +5,7 @@ from logzero import logger
 from neo.Settings import settings
 
 
-class DebugStorage():
+class DebugStorage:
 
     __instance = None