+++ neo/SmartContract/StateMachine.py
@@ -22,13 +22,7 @@ from neo.EventHub import SmartContractEvent
 
 
 class StateMachine(StateReader):
-
-    _accounts = None
     _validators = None
-    _assets = None
-    _contracts = None
-    _storages = None
-
     _wb = None
 
     _contracts_created = {}
@@ -66,18 +60,6 @@ class StateMachine(StateReader):
         self.Register("AntShares.Storage.Put", self.Storage_Put)
         self.Register("AntShares.Storage.Delete", self.Storage_Delete)
 
-    def CheckStorageContext(self, context):
-        if context is None:
-            return False
-
-        contract = self._contracts.TryGet(context.ScriptHash.ToBytes())
-
-        if contract is not None:
-            if contract.HasStorage:
-                return True
-
-        return False
-
     def ExecutionCompleted(self, engine, success, error=None):
 
         # commit storages right away
@@ -107,7 +89,6 @@ class StateMachine(StateReader):
         if account:
             engine.EvaluationStack.PushT(StackItem.FromInterface(account))
         else:
-            print("no account")
             engine.EvaluationStack.PushT(False)
 
         return True
@@ -192,7 +173,6 @@ class StateMachine(StateReader):
                 asset_type == AssetType.DutyFlag or \
                 asset_type == AssetType.GoverningToken or \
                 asset_type == AssetType.UtilityToken:
-
             return False
 
         if len(engine.EvaluationStack.Peek().GetByteArray()) > 1024:
@@ -325,7 +305,6 @@ class StateMachine(StateReader):
         contract = self._contracts.TryGet(hash.ToBytes())
 
         if contract is None:
-
             code = FunctionCode(script=script, param_list=param_list, return_type=return_type, contract_properties=contract_properties)
 
             contract = ContractState(code, contract_properties, name, code_version, author, email, description)
@@ -399,7 +378,6 @@ class StateMachine(StateReader):
             if contract.HasStorage:
 
                 for pair in self._storages.Find(engine.CurrentContext.ScriptHash()):
-
                     key = StorageKey(script_hash=hash, key=pair.Key.Key)
                     item = StorageItem(pair.Value.Value)
                     self._storages.Add(key, item)
@@ -425,7 +403,6 @@ class StateMachine(StateReader):
             created = self._contracts_created[shash.ToBytes()]
 
             if created == UInt160(data=engine.CurrentContext.ScriptHash()):
-
                 context = StorageContext(script_hash=shash)
                 engine.EvaluationStack.PushT(StackItem.FromInterface(context))
 
@@ -445,7 +422,6 @@ class StateMachine(StateReader):
             if contract.HasStorage:
 
                 for pair in self._storages.Find(hash.ToBytes()):
-
                     self._storages.Remove(pair.Key)
 
         self.events_to_dispatch.append(
@@ -470,7 +446,7 @@ class StateMachine(StateReader):
 
         key = engine.EvaluationStack.Pop().GetByteArray()
         storage_key = StorageKey(script_hash=context.ScriptHash, key=key)
-        item = self._storages.TryGet(storage_key.GetHashCodeBytes())
+        item = self._storages.TryGet(storage_key.ToArray())
 
         keystr = key
 
@@ -523,7 +499,7 @@ class StateMachine(StateReader):
 
         new_item = StorageItem(value=value)
         storage_key = StorageKey(script_hash=context.ScriptHash, key=key)
-        item = self._storages.GetOrAdd(storage_key.GetHashCodeBytes(), new_item)
+        item = self._storages.GetOrAdd(storage_key.ToArray(), new_item)
 
         keystr = key
         valStr = bytearray(item.Value)
@@ -563,6 +539,6 @@ class StateMachine(StateReader):
                                                               engine.ScriptContainer.Hash if engine.ScriptContainer else None,
                                                               test_mode=engine.testMode))
 
-        self._storages.Remove(storage_key.GetHashCodeBytes())
+        self._storages.Remove(storage_key.ToArray())
 
         return True