+++ CHANGELOG.rst
@@ -3,6 +3,21 @@ Changelog
 
 All notable changes to this project are documented in this file.
 
+[0.6.9] 2018-04-30
+-----------------------
+- alter logging
+- fix issue with dispatching transfer events when ``from_addr`` is ``False``
+- add TPS monitor to ``prompt`` ``state`` command
+- add check for db schema changes
+- add support for ``StorageIterator`` and ``Storage.Find`` from smart contracts
+- update to ``neocore==0.4.6``
+- update ``VERIFY`` implementation to match C#
+- add check for block tx length when adding to blockchain to address issues `#302 <https://github.com/CityOfZion/neo-python/issues/302>`_ and `#360 <https://github.com/CityOfZion/neo-python/issues/360>`_
+- add smart contract storage searching using a prefix.
+- source code cleanup of imports and class declarations
+- update to ``neo-boa==0.4.2``
+
+
 [0.6.8] 2018-04-26
 -----------------------
 - add ``ServiceEnabled`` boolean to settings to determine whether nodes should send other nodes blocks
@@ -18,9 +33,11 @@ All notable changes to this project are documented in this file.
 - fix asset amount rounding for very small amounts
 - fix storage commit routine for failed contract executions
 
+
 [0.6.7] 2018-04-06
 -----------------------
 - Update all the requirements
+- Networking changes
 - added ``--maxpeers`` option for ``np-prompt`` and ``np-api-server``.  This allows p2p discovery of new nodes up to the value specified
 - added ``--host`` option for ``np-api-server`` in order to specify a hostname for the server
 - added more testing for ``neo.Network`` module